EWD-22601 - Process server errors when user try to update deleted objective/experience/question Error on server when user try to delete objective that is related to experience
Former-commit-id: f91d6fce693481437b252c59690b73ea1c840711

+++ sources/easygenerator.Web/Controllers/Api/ObjectiveController.cs
@@ -79,6 +79,11 @@ namespace easygenerator.Web.Controllers.Api
         [HttpPost]
         public ActionResult Delete(Objective objective)
         {
+            if (objective.Experiences.Any())
+            {
+                return JsonLocalizableError(Constants.Errors.ObjectiveCannotBeDeleted,
+                                            Constants.Errors.ObjectiveCannotBeDeletedResourceKey);
+            }
             _repository.Remove(objective);
 
             return JsonSuccess();
